Error handling fixes to deal with missing pgbench or gnuplot.
The existing test to detect pgbench missing didn't actually work,
as it was trusting a test of $PGBENCH similar to this example:
FOO=""; if [ -f $FOO ]; then   echo "FOO appears to be a file"; fi

which doesn't quite work when the tested variable is an empty string.

Added some similar error handling to work around gnuplot not being
present without spewing a bunch of cryptic errors.