Rework reporting to better allow mixing test scripts in a set. Ordering now is usually set,script,scale,clients,rate_limit,test
Originally I expected every test in a set would use the same
test script.  That turned out to be a bad assumption.  Some of
the reports already handled that properly without me noticing.

To improve, I made a pass through every report looking for
groupings that did not include the script.  A few of these
intentionally use a different order, like report_latest and
report_orderbytest.