Add per-client rate limit CLIENTRATES that takes a target list. This is an alternative to setting an all client rate limit.
If you specify both SETRATES and CLIENTRATES the per-client value
will be used.

The logic in runset to implement this was optimized for a small diff
and is harder to follow now.  A later code cleanup might invert the
tests for empty strings (-z) as one way to clean up the implementation.