Cleanup the drop caches and benchmark tune scripts. System tool paths are now completely hard coded.
Where I tested, it wasn't reliable to expect the target sudo
environment to have both /usr/bin/ and /usr/sbin/ in its PATH.

Not sure if these tools are somewhere else in other distributions.
Testing so far has all been Debian/Ubuntu.

If it turns out these tools live in other directories in other
Linux distributions, will probably have to make a list and try
each.  Probably better to use a short hard coded list than
checking the PATH.  Given this is running as root, I'd prefer
not to have to trust that the path is free of shenanigans.